+++ .secrets.baseline
@@ -3,7 +3,7 @@
     "files": null,
     "lines": null
   },
-  "generated_at": "2020-10-08T14:00:58Z",
+  "generated_at": "2020-10-19T16:20:20Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
@@ -50,7 +50,7 @@
       {
         "hashed_secret": "a033a528b603fed46f861d4b3542c417b99d41c8",
         "is_verified": false,
-        "line_number": 66,
+        "line_number": 77,
         "type": "Secret Keyword"
       }
     ]
