+++ etl1/Dockerfile
@@ -1,4 +1,4 @@
-FROM python:3.7-slim
+FROM python:3.8-slim
 RUN  apt-get update
 RUN apt-get install -y build-essential
 RUN apt-get install -y libpq-dev
