+++ etl2/requirements.txt
@@ -1,0 +1,8 @@
+click
+sqlalchemy
+tabulate
+sqlalchemy
+psycopg2
+boto3
+pandas
+xlrd
