+++ terraform/ecs_etl2.tf
@@ -1,0 +1,66 @@
+resource "aws_ecs_task_definition" "etl2" {
+  family                   = "etl2-${terraform.workspace}"
+  requires_compatibilities = ["FARGATE"]
+  network_mode             = "awsvpc"
+  cpu                      = 2048
+  memory                   = 4096
+  container_definitions    = "[${local.etl2}]"
+  task_role_arn            = aws_iam_role.etl.arn
+  execution_role_arn       = aws_iam_role.execution_role.arn
+  tags = merge(local.default_tags,
+    { "Role" = "casrec-migration-${local.environment}" },
+  )
+}
+
+locals {
+  etl2 = jsonencode({
+    cpu       = 0,
+    essential = true,
+    image     = local.images.etl2,
+    name      = "etl2",
+    healthCheck = {
+      command     = ["CMD-SHELL", "echo hello || exit 1"],
+      startPeriod = 30,
+      interval    = 15,
+      timeout     = 10,
+      retries     = 3
+    },
+    logConfiguration = {
+      logDriver = "awslogs",
+      options = {
+        awslogs-group         = aws_cloudwatch_log_group.casrec_migration.name,
+        awslogs-region        = "eu-west-1",
+        awslogs-stream-prefix = "casrec-migration-${local.environment}"
+      }
+    },
+    secrets = [
+      {
+        name      = "DB_PASSWORD",
+        valueFrom = aws_secretsmanager_secret.casrec_migration.arn
+      },
+    ],
+    environment = [
+      {
+        name  = "DB_HOST",
+        value = aws_rds_cluster.cluster_serverless.endpoint
+      },
+      {
+        name  = "DB_PORT",
+        value = tostring(aws_rds_cluster.cluster_serverless.port)
+      },
+      {
+        name  = "DB_NAME",
+        value = aws_rds_cluster.cluster_serverless.database_name
+      },
+      {
+        name  = "DB_USER",
+        value = aws_rds_cluster.cluster_serverless.master_username
+      },
+      {
+        name  = "ENVIRONMENT",
+        value = terraform.workspace
+      }
+    ]
+  })
+}
+
